backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
litellm>=1.57.0
python-dotenv>=1.0.1

